Don't use threads (tasks) to add pids to cgroup
In xcgroup there was logic in the case cgroup.procs couldn't be read/write
then tasks file would be used instead. tasks contain a list of tids and pids
which means we can add threads to the cgroup, but that's not what we want.
We always want to deal with pids of processes, not threads, and cgroup will
take care of moving the entire process into the cgroup. If we cannot read or
find cgroup.procs we will not be able to read or write tasks file either.

There was also logic that if cgroup.procs were not found or we didn't have
permissions, then we would try to read /proc/<pid>/tasks directories and
would try to set anyway the cgroup.procs, which doesn't make much sense.

Note, there's one /proc/<pid>/tasks/<tid> directory per each thread of the
process.

Remove all this code and simplify things.