Add a workaround for cgroup slowness
In certain systems even if we don't see any pid in cgroup.procs file, it
doesn't mean the cgroup is ready to be deleted. This happens because the
kernel still keeps some internal references but already shows the file
as having no pids. In the kernel, css_has_online_children() call returns
EBUSY which is we can see sometimes.

If this happens, that means the cgroup system is slow. Even if we wait on
xcgroup_wait_pid_moved for not having pids in the cgroup.procs before trying
to remove the cgroup directories, we can still not be able to remove them.

So make it that if we receive a EBUSY, just sleep for a while and retry.

Bug 8911