Remove legacy check in xcgroup cpuset code
There was a legacy check in cpuset code for files named mems and cpuset,
withouth the prefix 'cpuset.' . This is not used anymore since all cgroup
even if mounted from /dev/cpuset, and even in Cray machines, do have the
prefix cpuset always set, so the cpuset.mems and cpuset.cpus files are the
only ones that can exist.

Remove unnecessary checks, signatures and logic to deal with this probably
ancient naming.