Do not add slurmstepd to job cgroups
In the original code the slurmstepd was not added to any cgroup except for
the freezer unless CoreSpecLimit, CoreSpecCount or MemSpecLimit was set.
During the refactor, since the code was made generic, the proctrack logic
was extended to all the controllers, turning in the stepd process being
always added to the corresponding container.

Initially this seemed a desired logic, because for example PMI does store a
key-value database in the stepd part, so a user process could make stepd to
consume a considerable amount of memory, so constraining this was a good
idea.

The issue is that adding the stepd to any cgroup means adding more logic
to add/remove the pid at the correct moment, and this was out of the scope
of this refactor. Also, it could cause undesired effects since the stepd
would now be constrained by job request, for example if a job just required
10MiB that could make stepd to crash.

This patch reverts this new logic, so the stepd forked by slurmd will reside
as before in the same cgroup than slurmd is, and will expect the admin to
set CoreSpecCount, CoreSpecLimit or MemSpecLimit in slurm.conf if he really
wants to control its memory or cpu resources.

We also avoid putting the stepd into the freezer container now. It seemed
not to buy anything, and all that provided is the need of logic to add and
remove the corresponding pid when the stepd cgroup was destroyed.

Bug 9254