Rename slurm_cgroup_conf_* functions
We do not want to have functions named slurm_* which causes them to be
exposed to the API. Instead we want to explictly set the aliases. So, rename
slurm_cgroup_conf_init and slurm_cgroup_conf_destroy, but keep the conf
struct with the slurm_ prefix and add the necessary aliases.

The only one which will use this aliases for the moment is pam_slurm_adopt.

Bug 11942