Do not set memory.force_empty anymore on jobacctgather and system cg
Since it was created, the jobacct gather cgroup plugin was setting the
force_empty parameter in the memory cgroup before destroying it. This
parameter was added in the kernel with the objective of reclaiming cgroup
file caches before deleting a cgroup. Doing that, one would ensure that
these out-of-use page caches were freed immediately instead of being kept
charged to the parent and reclaimed afterwards when memory pressure happens.

The documentation of cgroups says that the typical use case for this
interface is before calling rmdir(), but it is not clear which benefit does
it provide to Slurm. Moreover there's an important performance penalty when
terminating a job (~2s) which is undesirable. That might be a reason where
jobacctgather/cgroup was always considerably slower.

Finally this setting was only introduced in jobacctgather, but not in the
task plugin which also does use the memory controller.

The system cgroup also had this parameter which for consistency and because
we don't have a clear idea of the benefits, we are removing it too.

Kernel cgroup documentation:
5.1 force_empty
memory.force_empty interface is provided to make cgroup's memory usage
empty. When writing anything to this:

the cgroup will be reclaimed and as many pages reclaimed as possible.

The typical use case for this interface is before calling rmdir().
Though rmdir() offlines memcg, but the memcg may still stay there due to
charged file caches. Some out-of-use page caches may keep charged until
memory pressure happens. If you want to avoid that, force_empty will be
useful.

Also, note that when memory.kmem.limit_in_bytes is set the charges due to
kernel pages will still be seen. This is not considered a failure and the
write will still return success. In this case, it is expected that
memory.kmem.usage_in_bytes == memory.usage_in_bytes.

About use_hierarchy, see Section 6.

Bug 9254